Devel

+++ core/src/position.hpp
@@ -1,7 +1,6 @@
 #pragma once
-#include <cstddef>
 
 struct Position {
-    size_t row;
-    size_t col;
+    int row;
+    int col;
 };
